\name{simecolModels-package}
\alias{simecolModels-package}
\alias{simecolModels}
\docType{package}
\title{Collection of Model Objects for the simecol-Package
}
\description{
This package is a growing collection of models implemented in R according to the
simecol specification.
}
\details{
\tabular{ll}{
Package: \tab simecolModels\cr
Type: \tab Package\cr
Version: \tab 0.2-2\cr
Date: \tab 2008-07-23\cr
License: \tab Gnu Public License 2.0 or above\cr
}
}

\note{
In contrast to the demo models of the \pkg{simecol} package, the model objects
of this package are not stored as data but are generated by calling functions.
}

\references{
Petzoldt, T. and K. Rinke (2007).  simecol: An
Object-Oriented Framework for Ecological
Modeling in R.  Journal of Statistical
Software, 22(9), 1--31.  URL
\url{http://www.jstatsoft.org/v22/i09/}.

Please respect the specific references on the help pages of the respective models.
}

\seealso{
\code{\link[simecol:simecol-package]{simecol-package}}
}
\examples{
## -- sourcecode examples ----------------------------
## open the directory with all sourcecode examples
\dontrun{
browseURL(paste(system.file(package="simecolModels"), "/examples", sep=""))
} 
}
\keyword{ package }